Major updates
Fixed some bugs.
Converted the screenshot to size 640x640 and developed a way to convert each screenshot in a collection of cells defined as arrays. Added the script to visualize the splitted screenshot.
Adding the structure to split images into train and test correctly to be converted into ImageFolder and DataLoader.