Major updates
Added the bleeding version of the images to deal with bad cropping in order to expose the model to such variations. Fixed the val_loss and accuracy counter as well as added the metrics to evaluate the models performance. All of these changing are located inside the unico_test_set, which will most likely be the final version.